third_party/blink/web_tests/external/wpt/urlpattern/urlpattern.https.any.serviceworker-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Pattern: [{"protocol":"http","port":"80 "}] Inputs: [{"protocol":"http","port":"80"}]
  Failed to construct 'URLPattern': Invalid port pattern '80 '. Invalid port pattern '80 '.
[FAIL] Pattern: [{"port":"80"}] Inputs: [{"port":"8\\t0"}]
  assert_equals: test() result expected true but got false
[FAIL] Pattern: [{"port":"80"}] Inputs: [{"port":"80x"}]
  assert_equals: test() result expected true but got false
[FAIL] Pattern: [{"port":"80"}] Inputs: [{"port":"80?x"}]
  assert_equals: test() result expected true but got false
[FAIL] Pattern: [{"port":"80"}] Inputs: [{"port":"80\\\\x"}]
  assert_equals: test() result expected true but got false
[FAIL] Pattern: ["https://{sub.}?example{.com/}foo"] Inputs: ["https://example.com/foo"]
  Failed to construct 'URLPattern': Invalid hostname pattern '{sub.}?example{.com/}foo'. Invalid hostname pattern 'example.com/foo'.
[FAIL] Pattern: [{"hostname":"bad#hostname"}] Inputs: [{"hostname":"bad"}]
  Failed to construct 'URLPattern': Invalid hostname pattern 'bad#hostname'. Invalid hostname pattern 'bad#hostname'.
[FAIL] Pattern: [{"hostname":"bad/hostname"}] Inputs: [{"hostname":"bad"}]
  Failed to construct 'URLPattern': Invalid hostname pattern 'bad/hostname'. Invalid hostname pattern 'bad/hostname'.
[FAIL] Pattern: [{"hostname":"bad\\\\\\\\hostname"}] Inputs: [{"hostname":"badhostname"}]
  Failed to construct 'URLPattern': Invalid hostname pattern 'bad\\\\hostname'. Invalid hostname pattern 'bad\\hostname'.
[FAIL] Pattern: [{"hostname":"bad\\nhostname"}] Inputs: [{"hostname":"badhostname"}]
  Failed to construct 'URLPattern': Invalid hostname pattern 'bad\nhostname'. Invalid hostname pattern 'bad\nhostname'.
[FAIL] Pattern: [{"hostname":"bad\\rhostname"}] Inputs: [{"hostname":"badhostname"}]
  Failed to construct 'URLPattern': Invalid hostname pattern 'bad\rhostname'. Invalid hostname pattern 'bad\rhostname'.
[FAIL] Pattern: [{"hostname":"bad\\thostname"}] Inputs: [{"hostname":"badhostname"}]
  Failed to construct 'URLPattern': Invalid hostname pattern 'bad	hostname'. Invalid hostname pattern 'bad	hostname'.
Harness: the test ran to completion.
